+++ src/features/auth/hooks/useLostPasswordForm.ts
@@ -3,11 +3,11 @@ import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation } from '@tanstack/react-query';
 
-import { lostPassword } from '@server/auth/lostPassword';
 import type { FormNotification } from '@typings/forms/notification';
 
 import { LostPasswordFormSchema } from '../schemas/lostPassword.schema';
 import type { LostPasswordForm } from '../schemas/lostPassword.schema';
+import { lostPassword } from '../server/lostPassword';
 
 export default function useLostPasswordForm() {
   const [lostPasswordFormNotification, setLostPasswordFormNotification] = useState<FormNotification>({